call signMessage
