+++ apps/wallet-connect/package.json
@@ -4,9 +4,11 @@
   "private": true,
   "homepage": "./",
   "dependencies": {
+    "@ethersproject/abi": "^5.0.10",
     "@gnosis.pm/safe-react-components": "https://github.com/gnosis/safe-react-components.git#ff29c3c",
     "@walletconnect/client": "^1.3.3",
     "date-fns": "^2.16.1",
+    "ethereumjs-abi": "^0.6.8",
     "jsqr": "^1.3.1"
   },
   "scripts": {
@@ -28,5 +30,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/ethereumjs-abi": "^0.6.3"
   }
 }
